Basic Camera AI
First off note that the project file has been updated to VS2010 and most
likely won't open in 2008 anymore. The project has also been tweaked to
compile with XNA 4.0 If you need the 2008 file it should be in a Backup
folder.

All camera logic has been removed from the Game1 class and moved to the
Camera class. Basic Camera AI has been implemented. Because of this the
keyboard controls for the camera has been disabled. The code is still
there just commented out.