Map is now it's own class
Removed all the Map and Tile logic from the main game loop and put it in
it's own class. Tiles have been made a separate class as well since maps
and other objects need to interact with this. Both classes are still a
little rough. The variables need some fine-tuning as many are public and
shouldn't be. I tried having a public get and private set but it wasn't
working so I just made the public for now.