Hero refactored to Player. Small collision change
I made it so that the top collision will flip-flop sides depending on
your direction so the point is always above the pony's head.

We should add another point right above the flank so they can't jump
through a tile.

Also the collision points on the feet should be moved closer together.
It's possible to fall through a single tile if you stand on it just
right.