Lots of stuff, read description
[07/15/2012][1] Change Notes :

I changed Player back to Hero, and instead created a new Player class
that will store an array of playable heroes.

I added the ability to regain health by pressing the insert key (1 point
per press). This is just for testing and debugging. And in case I havent
mentioned it already, the delete key damages the hero (1 point per
press)

I altered the Animation class so that it is more self-contained.

I created 2 new sprite sheets for the map's terrain. One is a normal set
of tiles, and the other is for debugging. In the debugging sprite sheet,
all the tiles have a 1 pixel border around them. I also modified the Map
and Tiles classes so that they will use sprite sheets instead of
individual tile textures.

I added the toggleDebug() method to Game1. This method will toggle the
debug variable on itself as well as all other objects that have debug
variables, which right now, is only Hero. It also changes the map's
sprite sheet to the correct one. As more debug variables are added to
classes, this method will need to be modified to toggle those as well.
By default, debug is off. Pressing tilde now toggles the debug setting.

I fixed the collision problems for the most part, and I added a new top
point. we will probably have to rework them again, but for right now,
they do the job well enough.

The healing and damage delays in the KillableCreature class are now
functional, thanks to a function the tickTimers().

I also cleaned out some texture that were no longer needed
[TODO:] changing the collision points to be an offset value from the
entity's position. This means we will need to change the way
calculations are done on them.

There should be one more commit after this with some major things added