collision fixed
[07/15/2012][2] Change Notes :

Changed the collision points to be offsets of the position. I made
custom vector2 variables to that return the real world coodianates of
the collision points. I made the neccisary changes to the collision
calculations and now its working much better than before.

I changed the map creation so that it is now inside the Testlvl class.

I find myself having to hardcode far too many things, so I will probably
implement an XML seializer to automate everything